#122 support join on constructor injected path

+++ src/main/java/org/sfm/csv/impl/cellreader/ByteCellValueReaderUnbox.java
@@ -13,7 +13,7 @@ public class ByteCellValueReaderUnbox implements ByteCellValueReader {
 
     @Override
     public byte readByte(char[] chars, int offset, int length, ParsingContext parsingContext) {
-        return read(chars, offset, length, parsingContext).byteValue();
+        return read(chars, offset, length, parsingContext);
     }
 
     @Override
