+++ src/main/java/org/sfm/csv/impl/cellreader/IntegerCellValueReaderUnbox.java
@@ -13,7 +13,7 @@ public class IntegerCellValueReaderUnbox implements IntegerCellValueReader {
 
     @Override
     public int readInt(char[] chars, int offset, int length, ParsingContext parsingContext) {
-        return read(chars, offset, length, parsingContext).intValue();
+        return read(chars, offset, length, parsingContext);
     }
 
     @Override
